finish to simplify Java entity tests
