+++ kotysa-spring-jdbc/src/test/kotlin/org/ufoss/kotysa/spring/jdbc/postgresql/SpringJdbcJavaEntityPostgresqlTest.kt
@@ -4,105 +4,17 @@
 
 package org.ufoss.kotysa.spring.jdbc.postgresql
 
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.getBean
 import org.springframework.jdbc.core.JdbcOperations
-import org.springframework.transaction.PlatformTransactionManager
-import org.springframework.transaction.support.TransactionTemplate
-import org.ufoss.kotysa.spring.jdbc.sqlClient
-import org.ufoss.kotysa.spring.jdbc.transaction.transactionalOp
+import org.ufoss.kotysa.spring.jdbc.JavaUserRepository
+import org.ufoss.kotysa.spring.jdbc.SpringJdbcJavaEntityTest
 import org.ufoss.kotysa.tables
-import org.ufoss.kotysa.test.*
+import org.ufoss.kotysa.test.JavaUser
 
 
-class SpringJdbcJavaEntityPostgresqlTest : AbstractSpringJdbcPostgresqlTest<JavaUserPostgresqlRepository>() {
+class SpringJdbcJavaEntityPostgresqlTest :
+        AbstractSpringJdbcPostgresqlTest<JavaUserPostgresqlRepository>(), SpringJdbcJavaEntityTest<JavaUserPostgresqlRepository> {
     override val context = startContext<JavaUserPostgresqlRepository>()
-
     override val repository = getContextRepository<JavaUserPostgresqlRepository>()
-    private val transactionManager = context.getBean<PlatformTransactionManager>()
-    private val operator = TransactionTemplate(transactionManager).transactionalOp()
-
-    @Test
-    fun `Verify selectAll returns all users`() {
-        assertThat(repository.selectAll())
-                .hasSize(2)
-                .containsExactlyInAnyOrder(javaJdoe, javaBboss)
-    }
-
-    @Test
-    fun `Verify selectFirstByFirstame finds John`() {
-        assertThat(repository.selectFirstByFirstame("John"))
-                .isEqualTo(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectFirstByFirstame finds no Unknown`() {
-        assertThat(repository.selectFirstByFirstame("Unknown"))
-                .isNull()
-    }
-
-    @Test
-    fun `Verify selectByAlias1 finds TheBoss`() {
-        assertThat(repository.selectByAlias1("TheBoss"))
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaBboss)
-    }
-
-    @Test
-    fun `Verify selectByAlias2 finds TheBoss`() {
-        assertThat(repository.selectByAlias2("TheBoss"))
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaBboss)
-    }
-
-    @Test
-    fun `Verify selectByAlias3 finds TheBoss`() {
-        assertThat(repository.selectByAlias3("TheBoss"))
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaBboss)
-    }
-
-    @Test
-    fun `Verify selectByAlias1 with null alias finds John`() {
-        assertThat(repository.selectByAlias1(null))
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectAllByAlias2 with null alias finds John`() {
-        assertThat(repository.selectByAlias2(null))
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectByAlias3 with null alias finds John`() {
-        assertThat(repository.selectByAlias3(null))
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectAllMappedToDto does the mapping`() {
-        assertThat(repository.selectAllMappedToDto())
-                .hasSize(2)
-                .containsExactlyInAnyOrder(
-                        UserDto("John Doe", null),
-                        UserDto("Big Boss", "TheBoss"))
-    }
-
-    @Test
-    fun `Verify deleteAll works correctly`() {
-        operator.execute { transaction ->
-            transaction.setRollbackOnly()
-            assertThat(repository.deleteAll())
-                    .isEqualTo(2)
-            assertThat(repository.selectAll())
-                    .isEmpty()
-        }
-    }
 }
 
 private val tables =
@@ -111,12 +23,16 @@ private val tables =
                 name = "java_users"
                 column { it[JavaUser::getLogin].varchar() }
                         .primaryKey()
-                column { it[JavaUser::getFirstname].varchar {
-                    name = "fname"
-                } }
-                column { it[JavaUser::getLastname].varchar {
-                    name = "lname"
-                } }
+                column {
+                    it[JavaUser::getFirstname].varchar {
+                        name = "fname"
+                    }
+                }
+                column {
+                    it[JavaUser::getLastname].varchar {
+                        name = "lname"
+                    }
+                }
                 column { it[JavaUser::isAdmin].boolean() }
                 column { it[JavaUser::getAlias1].varchar() }
                 column { it[JavaUser::getAlias2].varchar() }
@@ -125,46 +41,4 @@ private val tables =
         }
 
 
-class JavaUserPostgresqlRepository(client: JdbcOperations) : Repository {
-
-    private val sqlClient = client.sqlClient(tables)
-
-    override fun init() {
-        createTable()
-        insert()
-    }
-
-    override fun delete() {
-        deleteAll()
-    }
-
-    private fun createTable() = sqlClient.createTable<JavaUser>()
-
-    fun insert() = sqlClient.insert(javaJdoe, javaBboss)
-
-    fun deleteAll() = sqlClient.deleteAllFromTable<JavaUser>()
-
-    fun selectAll() = sqlClient.selectAll<JavaUser>()
-
-    fun selectFirstByFirstame(firstname: String) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getFirstname] eq firstname }
-            .fetchFirstOrNull()
-
-    fun selectByAlias1(alias: String?) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getAlias1] eq alias }
-            .fetchAll()
-
-    fun selectByAlias2(alias: String?) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getAlias2] eq alias }
-            .fetchAll()
-
-    fun selectByAlias3(alias: String?) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getAlias3] eq alias }
-            .fetchAll()
-
-    fun selectAllMappedToDto() =
-            sqlClient.select {
-                UserDto("${it[JavaUser::getFirstname]} ${it[JavaUser::getLastname]}",
-                        it[JavaUser::getAlias1])
-            }.fetchAll()
-}
+class JavaUserPostgresqlRepository(client: JdbcOperations) : JavaUserRepository(client, tables)
