+++ kotysa-spring-r2dbc/src/test/kotlin/org/ufoss/kotysa/r2dbc/SpringR2dbcJavaEntityTest.kt
@@ -2,91 +2,90 @@
  * This is free and unencumbered software released into the public domain, following <https://unlicense.org>
  */
 
-package org.ufoss.kotysa.spring.jdbc
+package org.ufoss.kotysa.r2dbc
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.getBean
 import org.springframework.context.ConfigurableApplicationContext
-import org.springframework.transaction.PlatformTransactionManager
-import org.springframework.transaction.support.TransactionTemplate
-import org.ufoss.kotysa.spring.jdbc.transaction.transactionalOp
+import org.springframework.transaction.reactive.TransactionalOperator
+import org.ufoss.kotysa.r2dbc.transaction.transactionalOp
 import org.ufoss.kotysa.test.UserDto
 import org.ufoss.kotysa.test.javaBboss
 import org.ufoss.kotysa.test.javaJdoe
+import reactor.kotlin.test.test
 
-interface SpringJdbcJavaEntityTest<T : JavaUserRepository> {
+interface SpringR2dbcJavaEntityTest<T : JavaUserRepository> {
 
     val repository: T
     val context: ConfigurableApplicationContext
 
-    private val transactionManager get() = context.getBean<PlatformTransactionManager>()
-    private val operator get() = TransactionTemplate(transactionManager).transactionalOp()
+    private val operator get() = context.getBean<TransactionalOperator>().transactionalOp()
 
     @Test
     fun `Verify selectAll returns all users`() {
-        assertThat(repository.selectAll())
+        assertThat(repository.selectAll().toIterable())
                 .hasSize(2)
                 .containsExactlyInAnyOrder(javaJdoe, javaBboss)
     }
 
     @Test
     fun `Verify selectFirstByFirstame finds John`() {
-        assertThat(repository.selectFirstByFirstame("John"))
+        assertThat(repository.selectFirstByFirstame("John").block())
                 .isEqualTo(javaJdoe)
     }
 
     @Test
     fun `Verify selectFirstByFirstame finds no Unknown`() {
-        assertThat(repository.selectFirstByFirstame("Unknown"))
+        assertThat(repository.selectFirstByFirstame("Unknown").block())
                 .isNull()
     }
 
     @Test
     fun `Verify selectByAlias1 finds TheBoss`() {
-        assertThat(repository.selectByAlias1("TheBoss"))
+        assertThat(repository.selectByAlias1("TheBoss").toIterable())
                 .hasSize(1)
                 .containsExactlyInAnyOrder(javaBboss)
     }
 
     @Test
     fun `Verify selectByAlias2 finds TheBoss`() {
-        assertThat(repository.selectByAlias2("TheBoss"))
+        assertThat(repository.selectByAlias2("TheBoss").toIterable())
                 .hasSize(1)
                 .containsExactlyInAnyOrder(javaBboss)
     }
 
     @Test
     fun `Verify selectByAlias3 finds TheBoss`() {
-        assertThat(repository.selectByAlias3("TheBoss"))
+        assertThat(repository.selectByAlias3("TheBoss").toIterable())
                 .hasSize(1)
                 .containsExactlyInAnyOrder(javaBboss)
     }
 
     @Test
     fun `Verify selectByAlias1 with null alias finds John`() {
-        assertThat(repository.selectByAlias1(null))
+        assertThat(repository.selectByAlias1(null).toIterable())
                 .hasSize(1)
                 .containsExactlyInAnyOrder(javaJdoe)
     }
 
     @Test
     fun `Verify selectAllByAlias2 with null alias finds John`() {
-        assertThat(repository.selectByAlias2(null))
+        assertThat(repository.selectByAlias2(null).toIterable())
                 .hasSize(1)
                 .containsExactlyInAnyOrder(javaJdoe)
     }
 
     @Test
     fun `Verify selectByAlias3 with null alias finds John`() {
-        assertThat(repository.selectByAlias3(null))
+        assertThat(repository.selectByAlias3(null).toIterable())
                 .hasSize(1)
                 .containsExactlyInAnyOrder(javaJdoe)
     }
 
     @Test
     fun `Verify selectAllMappedToDto does the mapping`() {
-        assertThat(repository.selectAllMappedToDto())
+        assertThat(repository.selectAllMappedToDto().toIterable())
                 .hasSize(2)
                 .containsExactlyInAnyOrder(
                         UserDto("John Doe", null),
@@ -97,10 +96,10 @@ interface SpringJdbcJavaEntityTest<T : JavaUserRepository> {
     fun `Verify deleteAll works correctly`() {
         operator.execute { transaction ->
             transaction.setRollbackOnly()
-            assertThat(repository.deleteAll())
-                    .isEqualTo(2)
-            assertThat(repository.selectAll())
-                    .isEmpty()
-        }
+            repository.deleteAll()
+                    .doOnNext { n -> assertThat(n).isEqualTo(2) }
+                    .thenMany(repository.selectAll())
+        }.test()
+                .verifyComplete()
     }
 }
