+++ kotysa-spring-r2dbc/src/test/kotlin/org/ufoss/kotysa/r2dbc/h2/R2DbcJavaEntityH2Test.kt
@@ -4,104 +4,17 @@
 
 package org.ufoss.kotysa.r2dbc.h2
 
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.getBean
 import org.springframework.r2dbc.core.DatabaseClient
-import org.springframework.transaction.reactive.TransactionalOperator
-import org.ufoss.kotysa.r2dbc.sqlClient
-import org.ufoss.kotysa.r2dbc.transaction.transactionalOp
+import org.ufoss.kotysa.r2dbc.JavaUserRepository
+import org.ufoss.kotysa.r2dbc.SpringR2dbcJavaEntityTest
 import org.ufoss.kotysa.tables
-import org.ufoss.kotysa.test.*
-import reactor.kotlin.test.test
+import org.ufoss.kotysa.test.JavaUser
 
 
-class R2DbcJavaEntityH2Test : AbstractR2dbcH2Test<JavaUserH2Repository>() {
+class R2DbcJavaEntityH2Test :
+        AbstractR2dbcH2Test<JavaUserH2Repository>(), SpringR2dbcJavaEntityTest<JavaUserH2Repository> {
     override val context = startContext<JavaUserH2Repository>()
-
     override val repository = getContextRepository<JavaUserH2Repository>()
-    private val operator = context.getBean<TransactionalOperator>().transactionalOp()
-
-    @Test
-    fun `Verify selectAll returns all users`() {
-        assertThat(repository.selectAll().toIterable())
-                .hasSize(2)
-                .containsExactlyInAnyOrder(javaJdoe, javaBboss)
-    }
-
-    @Test
-    fun `Verify selectFirstByFirstame finds John`() {
-        assertThat(repository.selectFirstByFirstame("John").block())
-                .isEqualTo(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectFirstByFirstame finds no Unknown`() {
-        assertThat(repository.selectFirstByFirstame("Unknown").block())
-                .isNull()
-    }
-
-    @Test
-    fun `Verify selectByAlias1 finds TheBoss`() {
-        assertThat(repository.selectByAlias1("TheBoss").toIterable())
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaBboss)
-    }
-
-    @Test
-    fun `Verify selectByAlias2 finds TheBoss`() {
-        assertThat(repository.selectByAlias2("TheBoss").toIterable())
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaBboss)
-    }
-
-    @Test
-    fun `Verify selectByAlias3 finds TheBoss`() {
-        assertThat(repository.selectByAlias3("TheBoss").toIterable())
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaBboss)
-    }
-
-    @Test
-    fun `Verify selectByAlias1 with null alias finds John`() {
-        assertThat(repository.selectByAlias1(null).toIterable())
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectAllByAlias2 with null alias finds John`() {
-        assertThat(repository.selectByAlias2(null).toIterable())
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectByAlias3 with null alias finds John`() {
-        assertThat(repository.selectByAlias3(null).toIterable())
-                .hasSize(1)
-                .containsExactlyInAnyOrder(javaJdoe)
-    }
-
-    @Test
-    fun `Verify selectAllMappedToDto does the mapping`() {
-        assertThat(repository.selectAllMappedToDto().toIterable())
-                .hasSize(2)
-                .containsExactlyInAnyOrder(
-                        UserDto("John Doe", null),
-                        UserDto("Big Boss", "TheBoss"))
-    }
-
-    @Test
-    fun `Verify deleteAll works correctly`() {
-        operator.execute { transaction ->
-            transaction.setRollbackOnly()
-            repository.deleteAll()
-                    .doOnNext { n -> assertThat(n).isEqualTo(2) }
-                    .thenMany(repository.selectAll())
-        }.test()
-                .verifyComplete()
-    }
 }
 
 private val tables =
@@ -110,12 +23,16 @@ private val tables =
                 name = "java_users"
                 column { it[JavaUser::getLogin].varchar() }
                         .primaryKey()
-                column { it[JavaUser::getFirstname].varchar {
-                    name = "fname"
-                } }
-                column { it[JavaUser::getLastname].varchar {
-                    name = "lname"
-                } }
+                column {
+                    it[JavaUser::getFirstname].varchar {
+                        name = "fname"
+                    }
+                }
+                column {
+                    it[JavaUser::getLastname].varchar {
+                        name = "lname"
+                    }
+                }
                 column { it[JavaUser::isAdmin].boolean() }
                 column { it[JavaUser::getAlias1].varchar() }
                 column { it[JavaUser::getAlias2].varchar() }
@@ -124,48 +41,4 @@ private val tables =
         }
 
 
-class JavaUserH2Repository(dbClient: DatabaseClient) : Repository {
-
-    private val sqlClient = dbClient.sqlClient(tables)
-
-    override fun init() {
-        createTable()
-                .then(insert())
-                .block()
-    }
-
-    override fun delete() {
-        deleteAll()
-                .block()
-    }
-
-    private fun createTable() = sqlClient.createTable<JavaUser>()
-
-    fun insert() = sqlClient.insert(javaJdoe, javaBboss)
-
-    fun deleteAll() = sqlClient.deleteAllFromTable<JavaUser>()
-
-    fun selectAll() = sqlClient.selectAll<JavaUser>()
-
-    fun selectFirstByFirstame(firstname: String) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getFirstname] eq firstname }
-            .fetchFirst()
-
-    fun selectByAlias1(alias: String?) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getAlias1] eq alias }
-            .fetchAll()
-
-    fun selectByAlias2(alias: String?) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getAlias2] eq alias }
-            .fetchAll()
-
-    fun selectByAlias3(alias: String?) = sqlClient.select<JavaUser>()
-            .where { it[JavaUser::getAlias3] eq alias }
-            .fetchAll()
-
-    fun selectAllMappedToDto() =
-            sqlClient.select {
-                UserDto("${it[JavaUser::getFirstname]} ${it[JavaUser::getLastname]}",
-                        it[JavaUser::getAlias1])
-            }.fetchAll()
-}
+class JavaUserH2Repository(client: DatabaseClient) : JavaUserRepository(client, tables)
